% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubChem.R
\name{cid2sdf_online}
\alias{cid2sdf_online}
\title{cid2sdf_online}
\usage{
cid2sdf_online(cids, existTest = FALSE)
}
\arguments{
\item{cids}{cids vector.}

\item{existTest}{Testing cids url exist.}
}
\value{
A SDFset.
}
\description{
cid2sdf_online
}
\examples{
cid2sdf_online(cids = c(333, 3040203, 40030312), existTest = FALSE)
}
