% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMapping.R
\name{idMapping}
\alias{idMapping}
\title{idMapping}
\usage{
idMapping(id, to = "HMDB", default = "CID", unique = TRUE)
}
\arguments{
\item{id}{id vector with database name.}

\item{to}{database name.}
}
\value{
A id vector.
}
\description{
This function rely on metaboliteIDmapping data.
DTXCID: Comptox Chemical Dashboard
CID: Pubchem
CAS: CAS Registry numbers
HMDB: Human Metabolome Database
ChEBI: Chemical Entities of Biological Interest
KEGG: KEGG Compounds
Drugbank: Drugbank
}
\examples{
idMapping(id = "C18707")
id <- c(idMappingTb$CAS[1:5], idMappingTb$DTXSID[6:10], idMappingTb$DTXCID[4:11], idMappingTb$CID[12:15], idMappingTb$KEGG[12:15], idMappingTb$HMDB[16:19], idMappingTb$Drugbank[20:25])
idTb <- idMapping(id = id)
id <- c(idMappingTb$CAS[1:1000])
idTb <- idMapping(id = id, to = "HMDB", unique = FALSE)
}
