% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB.R
\name{load_hmdbCmpTb}
\alias{load_hmdbCmpTb}
\title{load_hmdbCmpTb}
\usage{
load_hmdbCmpTb()
}
\value{
A hmdbCmpTb tibble.
}
\description{
Load Human Metabolome Database (Version: 5.0).
The data is stored in a tibble with 217920 small molecules information,
which was generated from the xml file in \url{https://hmdb.ca/downloads}.
}
\examples{
hmdbCmpTb <- load_hmdbCmpTb()
}
\author{
Barry Song
}
