% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{load_keggCmpTb}
\alias{load_keggCmpTb}
\title{load_keggCmpTb}
\usage{
load_keggCmpTb()
}
\value{
keggCmpTb.
}
\description{
Load the KEGG Compounds Database.
}
\examples{
keggCmpTb <- load_keggCmpTb()
}
\author{
Barry Song.
}
