% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMSD.R
\name{load_lipidmapsCmpTb}
\alias{load_lipidmapsCmpTb}
\title{load_lipidmapsCmpTb}
\usage{
load_lipidmapsCmpTb()
}
\value{
A lipidmapsCmpTb.
}
\description{
load_lipidmapsCmpTb
}
\examples{
lipidmapsCmpTb <- load_lipidmapsCmpTb()
}
