% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubChem.R
\name{massSearch_CID}
\alias{massSearch_CID}
\title{massSearch_CID}
\usage{
massSearch_CID(mass, tol_mass = 0.001)
}
\arguments{
\item{mass}{A numeric vector of length 1 or 2.}

\item{tol_mass}{Use tol_mass to compute a mass range when the mass vector length is 1.}
}
\value{
A character vector with cids.
}
\description{
Searching compounds with exact mass.
}
\examples{
massSearch_CID(mass = 433.323)
}
\author{
Barry Song
}
