% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smiles.R
\name{plot_smiles}
\alias{plot_smiles}
\title{plot_smiles}
\usage{
plot_smiles(smiles)
}
\arguments{
\item{smiles}{A named vector.}
}
\value{
pic.
}
\description{
Input a smiles vector return a molecule pic.
}
\examples{
plot_smiles(smiles = "C-C-C-C")
}
