% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB.R
\name{search_hmdb_online}
\alias{search_hmdb_online}
\title{search_hmdb_online}
\usage{
search_hmdb_online(id)
}
\arguments{
\item{id}{A id vector.}
}
\value{
A hmdbCmpTb tibble.
}
\description{
Search HMDB database online.
}
\examples{
hmdbCmpTb_test <- search_hmdb_online(id = c("HMDB0000001", "HMDB0000002"))
}
