% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubChem.R
\name{search_pubchem_online}
\alias{search_pubchem_online}
\title{search_pubchem_online}
\usage{
search_pubchem_online(
  inputs,
  type = c("cid", "name", "smiles", "inchi", "inchikey")[1],
  synonyms = FALSE
)
}
\arguments{
\item{inputs}{A character vector.}

\item{type}{Supports five search types, cid, name, smiles, inchi and inchikey.}

\item{synonyms}{Whether to return synonyms in the result.}
}
\value{
A tibble.
}
\description{
Search PubChem database online.
}
\examples{
search_pubchem_online(inputs = c("333", "4221"))
search_pubchem_online(inputs = c("Water"), type = "name")
search_pubchem_online(inputs = massSearch_CID(mass = 433.323, tol_mass = 0.0001), type = "cid")
}
